Move imageprovider package out of httpcontroller
