+++ internal/httpcontroller/init.go
@@ -5,7 +5,6 @@ import (
 	"html/template"
 	"io"
 	"log"
-	"sync"
 
 	"github.com/labstack/echo/v4"
 	"github.com/labstack/echo/v4/middleware"
@@ -13,7 +12,7 @@ import (
 
 	"github.com/tphakala/birdnet-go/internal/conf"
 	"github.com/tphakala/birdnet-go/internal/datastore"
-	"github.com/tphakala/birdnet-go/internal/httpcontroller/imageprovider"
+	"github.com/tphakala/birdnet-go/internal/imageprovider"
 	"github.com/tphakala/birdnet-go/internal/logger"
 )
 
@@ -37,20 +36,15 @@ type Server struct {
 }
 
 // New initializes a new HTTP server with given context and datastore.
-func New(settings *conf.Settings, dataStore datastore.Interface) *Server {
+func New(settings *conf.Settings, dataStore datastore.Interface, birdImageCache *imageprovider.BirdImageCache) *Server {
 	// Default port configuration
 	configureDefaultSettings(settings)
-
-	cache, err := imageprovider.CreateDefaultCache()
-	if err != nil {
-		log.Fatal(err)
-	}
 
 	s := &Server{
 		Echo:           echo.New(),
 		ds:             dataStore,
 		Settings:       settings,
-		BirdImageCache: cache,
+		BirdImageCache: birdImageCache,
 	}
 
 	// Server initialization
@@ -106,7 +100,6 @@ func (s *Server) initializeServer() {
 	s.initLogger()
 	s.configureMiddleware()
 	s.initRoutes()
-	go s.initBirdImageCache()
 }
 
 // handleServerError listens for server errors and handles them.
@@ -151,27 +144,4 @@ func (s *Server) initLogger() {
 			return nil
 		},
 	}))
-}
-
-func (s *Server) initBirdImageCache() {
-	speciesList, err := s.ds.GetAllDetectedSpecies()
-	if err != nil {
-		s.Echo.Logger.Fatal(err)
-	}
-
-	var wg sync.WaitGroup
-
-	for _, species := range speciesList {
-		wg.Add(1)
-
-		go func(speciesName string) {
-			defer wg.Done()
-			_, err := s.BirdImageCache.Get(speciesName)
-			if err != nil {
-				s.Echo.Logger.Error(err)
-			}
-		}(species.ScientificName)
-	}
-
-	wg.Wait()
 }
